"Chunks" kind of
Chunks store 4096 blocks in localised coordinates with each dimension being in range 0 to 15. Chunk dimensions are 0 to n

to get a global block position we use a transformation matrix on the chunk to account for the position of the block locally in the chunk, and the chunk position globally

Next on the list is to add greedy meshing by building up chunk meshes which are expanded by two blocks in each dimension to check face visibility and skip rendering if not visible.